Show the upload dialog once the upload is success
